Reorganize atoms, molecules, organisms
