+++ frontend/src/components/molecules/Avatar.tsx
