+++ frontend/src/components/organisms/EventDetails.tsx
