+++ frontend/src/components/templates/DefaultTemplate.tsx
@@ -1,5 +1,5 @@
 import React, { ReactNode } from "react";
-import NavBar from "@/components/molecules/NavBar";
+import NavBar from "@/components/organisms/NavBar";
 
 interface DefaultTemplateProps {
   children: ReactNode;
