+++ frontend/src/components/templates/WelcomeTemplate.tsx
@@ -1,5 +1,5 @@
 import React, { ReactNode } from "react";
-import NavBar from "@/components/molecules/NavBar";
+import NavBar from "@/components/organisms/NavBar";
 
 /** A WelcomeTemplate page */
 interface WelcomeTemplateProps {
