+++ frontend/src/pages/events/[eventid]/edit.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { useRouter } from "next/router";
-import EventForm from "@/components/molecules/EventForm";
+import EventForm from "@/components/organisms/EventForm";
 import CenteredTemplate from "@/components/templates/CenteredTemplate";
 
 /** An EditEvent page */
