+++ frontend/src/pages/events/view.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import ViewEvents from "../../components/molecules/ViewEvents";
+import ViewEvents from "../../components/organisms/ViewEvents";
 import DefaultTemplate from "../../components/templates/DefaultTemplate";
 
 /** A ViewEventsPage page */
