+++ frontend/src/pages/password/reset/[link].tsx
@@ -1,6 +1,6 @@
 import React from "react";
 // import { useRouter } from "next/router";
-import ResetPasswordForm from "@/components/molecules/ResetPasswordForm";
+import ResetPasswordForm from "@/components/organisms/ResetPasswordForm";
 import WelcomeTemplate from "@/components/templates/WelcomeTemplate";
 
 /** A ResetPassword page */
