+++ frontend/src/pages/profile.tsx
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from "react";
-import ProfileForm from "@/components/molecules/ProfileForm";
+import ProfileForm from "@/components/organisms/ProfileForm";
 import ProfileTemplate from "@/components/templates/ProfileTemplate";
 import Banner from "@/components/molecules/Banner";
 import { GetServerSideProps } from "next";
@@ -7,7 +7,7 @@ import { BASE_URL } from "@/utils/constants";
 import { onAuthStateChanged, getAuth } from "firebase/auth";
 import { auth } from "@/utils/firebase";
 import { useAuth } from "@/utils/AuthContext";
-import Avatar from "@/components/atoms/Avatar";
+import Avatar from "@/components/molecules/Avatar";
 
 type userData = {
   id: string;
