+++ frontend/src/pages/signup.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import WelcomeTemplate from "@/components/templates/WelcomeTemplate";
-import SignupForm from "@/components/molecules/SignupForm";
+import SignupForm from "@/components/organisms/SignupForm";
 import { useEffect } from "react";
 import { useRouter } from "next/router";
 import { useAuth } from "@/utils/AuthContext";
